refactor: JSONDecodable 타입 구조 변경
- 기존 JSONDecodable 의 protocol + extension 조합 해제.
- JSONDecodable 프로토콜을 채택한 DecodingManager 타입으로 변경.
- 불필요한 메서드 삭제 : readLocalFile(for:) throws -> Data
- 파일 이름 변경 : JSONDecodable.swift -> DecodingManager.swift
- APIGetCallable.swift 에서 callGetAPI() 메서드 내의 decodeManager.decodeJSON() 메서드 활용 방식 변경
  > 의존성 주입을 통한 메서드 호출방식으로 변경.
